% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_lambda.R
\name{choose_lambda}
\alias{choose_lambda}
\title{Extract results for a chosen value of \eqn{\lambda}}
\usage{
choose_lambda(x, lambda = 1)
}
\arguments{
\item{x}{An object of class \code{"bcthresh"} returned from
\code{\link{bcthresh}}.}

\item{lambda}{A numeric scalar.  Must be contained in \code{x$lambda}.}
}
\value{
An object of class \code{"ithresh"}.  See \code{\link{ithresh}}.
}
\description{
Extract from an object of class \code{"bcthresh"} returned from
\code{\link{bcthresh}} measures of predictive performance and
posterior samples obtained after a Box-Cox transformation of the data
using a particular value of \eqn{\lambda}.
}
\details{
The function simply extracts the relevant objects from \code{x}
  and recreates the transformed data, training and validation thresholds
  using the values of \code{lambda} and the normalizing constant
  stored in \code{x$lngm}.  The returned object has class \code{"ithresh"}
  and the same structure as an object returned from \code{\link{ithresh}},
  for which \code{plot}, \code{summary} and \code{predict} methods are
  available.
}
\examples{
probs <- seq(0, 0.9, 0.1)
lambda <- seq(1, 3, 0.5)

prior_args <- list(prior = "flatflat", bin_prior = "haldane",
                   h_prior = list(min_xi = -Inf))
gom_args <- list(data = gom, probs = probs, lambda = lambda)
gom_lambda <- do.call(bcthresh, c(gom_args, prior_args))

# lambda = 1 (the default)
res1 <- choose_lambda(gom_lambda)
plot(res1)
# lambda = 2
res2 <- choose_lambda(gom_lambda, lambda = 2)
plot(res2)
}
\seealso{
\code{\link{plot.ithresh}} for the S3 plot method for objects of
  class \code{ithresh}.

\code{\link{summary.ithresh}} Summarizing measures of threshold
  predictive performance.

\code{\link{predict.ithresh}} for predictive inference for the
  largest value observed in N years.
}
