% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshr.R
\docType{data}
\name{ns}
\alias{ns}
\title{Storm peak significant wave heights from the North Sea}
\format{A vector containing 628 observations.}
\source{
Oceanweather Inc. (1995) NEXT -- North Sea hindcast study.
}
\usage{
ns
}
\description{
A numeric vector containing 628 hindcasts of storm peak significant wave
heights, in metres, from 1964 to 1995 at an unnamed location in the North
Sea.
}
\references{
Northrop, P. J., N. Attalides, and P. Jonathan. (2017).
  Cross-Validatory Extreme Value Threshold Selection and Uncertainty with
  Application to Ocean Storm Severity. \emph{Journal of the Royal
  Statistical Society: Series C (Applied Statistics)}, \strong{66}(1),
  93-120.
  doi:\href{https://doi.org/10.1111/rssc.12159}{10.1111/rssc.12159}.
}
\keyword{datasets}
