% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthresh_methods.R
\name{plot.bcthresh}
\alias{plot.bcthresh}
\title{Plot method for objects of class "bcthresh"}
\usage{
\method{plot}{bcthresh}(x, which_v = 1,
  which_lambdas = 1:length(x$lambda), legend_pos = "bottom", ...)
}
\arguments{
\item{x}{an object inheriting from class "bcthresh", a result of a call to
\code{\link{bcthresh}}.}

\item{which_v}{A numeric scalar.  Specifies the validation threshold, that
is the components of \code{x$v_vec}, to use in the plot.}

\item{which_lambdas}{A numeric vector.  Specifies which values of
\eqn{\lambda}, that is, the components of \code{x$lambda}, to include in
the plot.  The default is to use all these values.}

\item{legend_pos}{The position of the legend specified using the argument
\code{x} in \code{\link[graphics]{legend}}.}

\item{...}{Additional graphical parameters to be passed to
\code{\link[graphics]{matplot}} and \code{\link[graphics]{legend}}.
The default setting plots solid lines (\code{lty = 1}) of width 2
(\code{lwd = 2}) using \code{col = 1:length(x$lambda)}.}
}
\value{
Nothing.
}
\description{
\code{plot} method for class "bcthresh".
}
\details{
Plots the measure of predictive performance against quantile
  of training threshold for different values of \eqn{\lambda}: those
  stored in \code{x$lambda}.  It is possible that a curve on the plot
  may be incomplete.  This indicates that, for a particular \eqn{\lambda}
  and threshold level, a measure of predictive performance is \code{-Inf}.
  This occurs when an observation in the data lies above the estimated upper
  end point of the predictive distribution produced when this observation is
  removed.
}
\section{Examples}{

See the examples in \code{\link{bcthresh}}.
}

