% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthreshpred_methods.R
\name{plot.bcthreshpred}
\alias{plot.bcthreshpred}
\title{Plot diagnostics an bcthreshpred object}
\usage{
\method{plot}{bcthreshpred}(x, lambda = x$lambda, n_years = if
  (length(lambda) == 1) x$n_years else x$n_years[1], legend_pos, ...)
}
\arguments{
\item{x}{an object of class "bthreshpred", a result of a call to
\code{\link{ithresh}}.}

\item{lambda}{A numeric vector.  Specifies which values of
\eqn{\lambda} to include in the plot.  These values must be present
in \code{x$lambda}.  The default is to use all these values.}

\item{n_years}{A numeric vector.  Specifies which values in
\code{x$n_years} to use for the time horizon.  If
\code{length(lambda) > 1} then \code{n_years} must have length 1.
The defaults are that \code{x$n_years[1]} is used if
\code{length(lambda) > 1} and \code{x$years} otherwise.}

\item{legend_pos}{The position of the legend specified using the argument
\code{x} in \code{\link[graphics]{legend}}.  If \code{legend_pos} is
missing then it is set internally, based on \code{x$type}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{matplot}}.}
}
\description{
\code{plot} method for class "bcthreshpred".  Produces plots to summarise
the predictive inferences made by \code{\link{predict.bcthresh}}.
}
