% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.ithresh}
\alias{plot.ithresh}
\title{Plot diagnostics for a thresh}
\usage{
\method{plot}{ithresh}(x, y, which_val = NULL, prob = TRUE,
  top_scale = TRUE, add_legend = FALSE, legend_pos = "topleft", ...)
}
\arguments{
\item{x}{an object of class "ithresh", a result of a call to
\code{\link{ithresh}}.}

\item{y}{Not used.}

\item{which_val}{A numeric vector specifying the validation thresholds, that
is the indices of the argument \code{v_vec} to \code{\link{ithresh}}.}

\item{prob}{A logical scalar.  If \code{TRUE} then the levels of thresholds
are represented by the proportion of observations that lie below a
threshold.  If \code{prob = FALSE} then the values of the thresholds are
used.}

\item{top_scale}{A logical scalar indicating Whether or not to add a scale
to the top horizontal axis.  If this is added it gives the threshold on
the scale not chosen by \code{prob}.}

\item{add_legend}{A logical scalar indicating whether or not to add a
legend to the plot.  If \code{method = "cv"} then the legend gives the
levels of the validation thresholds.}

\item{legend_pos}{The position of the legend (if required) specified using
the argument \code{x} in \code{\link[graphics]{legend}}.}

\item{...}{Additional arguments passed on to \code{\link[graphics]{matplot}}
and/or \code{\link[graphics]{legend}} and/or \code{\link[graphics]{axis}}.}
}
\description{
\code{plot} method for class "ithresh".
}
\details{
Add some details.
}
\examples{
\dontrun{
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
gom_cv <- ithresh(data = gom, u_vec = u_vec, n_v = 4)
plot(gom_cv, lwd = 2, add_legend = TRUE, legend_pos = "topleft")
mtext("significant wave height / m", side = 3, line = 2.5)
}
}
