% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.thresh}
\alias{plot.thresh}
\title{Plot diagnostics for a thresh}
\usage{
\method{plot}{thresh}(x, y, ...)
}
\arguments{
\item{x}{an object of class "thresh", a result of a call to
\code{\link{ithresh}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to \code{matplot}.}
}
\description{
\code{plot} method for class "thresh".
}
\details{
Add some details.
}
\examples{
# NAJ 2016
library(revdbayes)
data(gom)
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
v_vec <- quantile(gom, probs = c(0.8, 0.85, 0.9, 0.95))
naj_control <- list(prior_args = list(max_xi = 1))
gom_naj <- ithresh(data = gom, method = "naj", u_vec = u_vec, v_vec = v_vec)
plot(gom_naj)
}

