% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predict.bcthresh}
\alias{predict.bcthresh}
\title{Predictive inference for the largest value observed in N years.}
\usage{
\method{predict}{bcthresh}(object, lambda, ...)
}
\arguments{
\item{object}{An object of class \code{"bcthresh"}, a result of a call to
\code{\link{bcthresh}}.}

\item{lambda}{A numeric vector containing values of the Box-Cox
transformation parameter \eqn{\lambda} to use when performing predictive
inference.  These must be contained in \code{object$lambda}.}

\item{...}{Additional arguments to be passed to
\code{\link{predict.ithresh}}.}
}
\description{
\code{predict} method for class \code{"bcthresh"}.  Calls
\code{\link{predict.ithresh}} for each value of \eqn{\lambda} in
\code{lambda}.
}
