% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthresh_methods.R
\name{print.bcthresh}
\alias{print.bcthresh}
\title{Print method for objects of class "bcthresh"}
\usage{
\method{print}{bcthresh}(x, digits = 2, ...)
}
\arguments{
\item{x}{an object inheriting from class "bcthresh", a result of a call to
\code{\link{bcthresh}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{format}} and \code{\link[base:Round]{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class "bcthresh".
}
\details{
For each value in \code{x$lambda}, a matrix of the estimated
  threshold weights is printed.
  Each row gives the weights for each training threshold for a given
  validation threshold.  The row and column names are the approximate
  quantile levels of the thresholds.
}
