% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_simulation.R
\name{rbc}
\alias{rbc}
\title{Simulation with (Inverse) Box-Cox Transformation}
\usage{
rbc(n = 1, lambda = 1, sim_fn = stats::rexp, ...)
}
\arguments{
\item{n}{An integer scalar.  The sample size required.}

\item{lambda}{A numeric scalar.  The Box-Cox transformation parameter.}

\item{sim_fn}{An R function that simulates from the target probability
distribution.}

\item{...}{Further arguments to be passed to \code{sim_fn}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Random generation for a randm variable that has the following property:
a Box-Cox transformation of the variable has a specified target
distribution.
}
\details{
A sample of size \code{n} is simulated using the function supplied
  in \code{sim_fn}.  Then the inverse of the Box-Cox transformation is
  applied to each value in this sample.
}
\examples{
x <- rbc(n = 100, lambda = 2)
x <- rbc(n = 100, lambda = 2, sim_fn = revdbayes::rgp)
x <- rbc(n = 100, lambda = 2, sim_fn = revdbayes::rgp, shape = 0.1)
}
