% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthresh_methods.R
\name{summary.bcthresh}
\alias{summary.bcthresh}
\title{Summarizing measures of threshold predictive performance}
\usage{
\method{summary}{bcthresh}(object, digits = 2, ...)
}
\arguments{
\item{object}{an object inheriting from class "bcthresh", a result of a call
to \code{\link{bcthresh}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{format}} and \code{\link[base:Round]{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
Returns a numeric matrix with 6 columns and \code{n_v} rows,
  where \code{n_v} is an argument to \code{\link{ithresh}} that
  determines how many of the largest training thresholds are used
  a validation thresholds.  The first column is the value of \eqn{\lambda}.
  The remaining 5 columns are those returned from
  \code{\link{summary.ithresh}}.
}
\description{
\code{summary} method for class "bcthresh"
}
\section{Examples}{

See the examples in \code{\link{bcthresh}}.
}

