% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshr-internal.R
\name{threshr-internal}
\alias{threshr-internal}
\alias{bc}
\alias{inv_bc}
\alias{yj}
\alias{inv_yj}
\alias{diag_pos}
\title{Internal threshr functions}
\usage{
bc(x, lambda = 1, lambda_tol = 1/50, m = 4)

inv_bc(x, lambda = 1, lambda_tol = 1/50, m = 4)

yj(x, lambda = 1, lambda_tol = 1/50, m = 4)

inv_yj(x, lambda = 1, lambda_tol = 1/50, m = 4)

diag_pos(x)
}
\description{
Internal threshr functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
