% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshr.R
\docType{package}
\name{threshr}
\alias{threshr}
\alias{threshr-package}
\title{threshr: Threshold Selection and Uncertainty for Extreme Value Analysis}
\description{
Provides functions for the selection of extreme value threshold.
See the 'threshr' website for more information, documentation and examples.
}
\details{
The main function in the threshr package is

  See \code{vignette("threshr-vignette", package = "threshr")} for an
  overview of the package.
}
\references{
Northrop, P. J. (2016). revdbayes: Ratio-of-Uniforms Sampling
  for Bayesian Extreme Value Analysis. R package version 1.0.0.
  \url{https://cran.r-project.org/package=revdbayes}.
}
\seealso{
\code{\link[revdbayes]{set_prior}} and
  \code{\link[revdbayes]{rpost}}: functions from the
  \code{\link{revdbayes}} package used to set an extreme value prior density
  and to sample from a extreme value posterior density respectively.
}

