% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithresh.R
\name{ithresh}
\alias{ithresh}
\title{Threshold selection in the i.i.d. case (peaks over threshold)}
\usage{
ithresh(data, u_vec, n_v = 1, use_rcpp = TRUE, ...)
}
\arguments{
\item{data}{A numeric vector of observations.}

\item{u_vec}{A numeric vector. A vector of \emph{training} thresholds
at which inferences are made from the GP model.  Any duplicated values
will be removed. These could be set at sample quantiles of \code{data}
using \code{\link[stats]{quantile}}.}

\item{n_v}{A numeric scalar.
The \code{n_v} largest values in \code{u_vec} will be used to quantify
the predictive performance of the GP models fitted at the thresholds
in \code{u_vec}.}

\item{use_rcpp}{A logical scalar.  If TRUE (the default) use the
revdbayes function \code{\link[revdbayes]{rpost_rcpp}} for posterior
simulation.  Otherwise, we use \code{\link[revdbayes]{rpost}}.}

\item{...}{Further (optional) arguments to be passed to the
  \code{\link[revdbayes]{revdbayes}} function
  \code{\link[revdbayes]{rpost_rcpp}} (or \code{\link[revdbayes]{rpost}}).
  In particular:
\itemize{
  \item {\code{n}} {The size of the posterior sample used to perform
    predictive inference.  Default: \code{n = 1000}.}
  \item {\code{prior}} {A prior for the GP parameters, set using
    \code{\link[revdbayes]{set_prior}}.  Default: \code{prior = "flat"}
    with \code{min_xi = -1}.}
  \item {\code{h_prior}} {A list of further arguments (hyperparameters)
    for the GP prior specified in \code{prior}.}
  \item {\code{bin_prior}} {A prior for the threshold exceedance
    probability \eqn{p}, set using \code{\link[revdbayes]{set_bin_prior}}.
    Default: \code{prior = "jeffreys"}.}
  \item {\code{h_bin_prior}} {A list of further arguments (hyperparameters)
    for the binomial prior specified in \code{bin_prior}.}
}}
}
\value{
An object (list) of class \code{"thresh"}.
}
\description{
Produces diagnostic plots to assist in the selection of an extreme value
threshold in the case where the data can be treated as independent and
identically distributed (i.i.d.) observations.  For example, it could be
that these observations are the cluster maxima resulting from the
declustering of time series data.  The diagnostic plots are based on
making inferences from a Generalised Pareto (GP) model for threshold
excesses.
}
\details{
Add some details.
\emph{Northrop, Attalides and Jonathan (2016)}:
See the threshr vignette for further details and examples.
}
\examples{
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))

# Cross-validation approach
gom_cv <- ithresh(data = gom, u_vec = u_vec, n_v = 4)
plot(gom_cv)

u_vec <- quantile(ns, probs = seq(0, 0.95, by = 0.05))
# Cross-validation approach
cv_control <- list(prior = "mdi", h_prior = list(a = 0.6))
ns_cv <- ithresh(data = ns, u_vec = u_vec, n_v = 4,
  prior = "mdi", h_prior = list(a = 0.6))
plot(ns_cv)
}
\references{
Northrop, P.J. and Attalides, N. (2016) Posterior propriety in
  Bayesian extreme value analyses using reference priors
  \emph{Statistica Sinica}, \strong{26}(2), 721--743
  \url{http://dx.doi.org/10.5705/ss.2014.034}.

Northrop, P. J., Attalides, N. and Jonathan, P. (2017)
  Cross-validatory extreme value threshold selection and uncertainty
  with application to ocean storm severity.
  \emph{Journal of the Royal Statistical Society Series C: Applied
  Statistics}, \strong{66}(1), 93-120.
  \url{http://dx.doi.org/10.1111/rssc.12159}
}
\seealso{
\code{\link[revdbayes]{rpost}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of the arguments
  that can be passed to
  \code{\link[revdbayes]{rpost_rcpp}}/\code{\link[revdbayes]{rpost}}.

\code{\link[revdbayes]{set_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for GP parameters.

\code{\link[revdbayes]{set_bin_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for the exceedance probability \eqn{p}.

\code{\link[stats]{quantile}}.
}
