% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithresh.R
\name{ithresh}
\alias{ithresh}
\title{Threshold selection in the i.i.d. case (peaks over threshold).}
\usage{
ithresh(data, method = c("stability", "naj", "nc", "wadsworth"), u_vec,
  v_vec = max(u_vec), naj_control = list(), ...)
}
\arguments{
\item{data}{A numeric vector of observations.}

\item{method}{A character scalar.  The method to be used.
\itemize{
  \item {"stability"} {A parameter estimate stability plot for the GP
    shape parameter \eqn{\xi}}
  \item {"naj"} {Northrop, Attalides and Jonathan (2016)}
  \item {"nc"} {Northrop and Coleman (2014)}
  \item {"wadsworth"} {Wadsworth (2015)}
}}

\item{u_vec}{A numeric vector. A vector of \emph{training} thresholds
at which inferences are made from the GP model.}

\item{v_vec}{A numeric vector. Relevant only if \code{method = "naj"}.
A vector of \emph{validation} thresholds used to quantify the predictive
performance of the GP models fitted at the thresholds in \code{u_vec}.}

\item{naj_control}{A list of (optional) arguments for use if
  \code{method = "naj"}.  In particular:
\itemize{
  \item {\code{prior}} {A prior for the GP parameters, set using
    \code{\link[revdbayes]{set_prior}}.  Default: \code{prior = "flat"}.
    with \code{min_xi = -1}.}
  \item {\code{h_prior}} {A list of further arguments (hyperparameters)
    for the GP prior specified in \code{prior}.}
  \item {\code{bin_prior}} {A prior for the threshold exceedance
    probability \eqn{p}, set using \code{\link[revdbayes]{set_bin_prior}}.
    Default: \code{prior = "jeffreys"}.}
  \item {\code{h_bin_prior}} {A list of further arguments (hyperparameters)
    for the binomial prior specified in \code{bin_prior}.}
  \item {\code{n}} {The size of the posterior sample used to perform
    predictive inference.  Default: \code{n = 1000}.}
  \item {...} {Further arguments to \code{prior} or \code{bin_prior}.}
}}

\item{...}{Further argments to be passed to}
}
\value{
An object (list) of class \code{"thresh"}.
}
\description{
Produces diagnostic plots to assist in the selection of an extreme value
threshold in the case where the data can be treated as independent and
identically distributed (i.i.d.) observations.  For example, it could be
that these observations are the cluster maxima resulting from the
declustering of time series data.  The diagnostic plots are based on
making inferences from a Generalised Pareto (GP) model for threshold
excesses.
}
\details{
Add some details.
\emph{parameter estimate stability plot}:
\emph{Northrop, Attalides and Jonathan (2016)}:
\emph{Northrop and Coleman (2014)}:
\emph{Wadsworth (2015)}:
See the threshr vignette for further details and examples.
}
\examples{
# NAJ 2016
library(revdbayes)
data(gom)
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
v_vec <- quantile(gom, probs = c(0.8, 0.85, 0.9, 0.95))
naj_control <- list(prior_args = list(max_xi = 1))
gom_naj <- ithresh(data = gom, method = "naj", u_vec = u_vec, v_vec = v_vec)
}
\seealso{
\code{\link[revdbayes]{rpost}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of the arguments
  that can be passed to \code{rpost}.

\code{\link[revdbayes]{set_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for GP parameters.

\code{\link[revdbayes]{set_bin_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for the exceedance probability \eqn{p}.
}

