% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithresh.R
\name{ithresh}
\alias{ithresh}
\title{Threshold selection in the i.i.d. case (peaks over threshold)}
\usage{
ithresh(data, u_vec, n_v = 1, use_rcpp = TRUE, ...)
}
\arguments{
\item{data}{A numeric vector of observations.}

\item{u_vec}{A numeric vector. A vector of \emph{training} thresholds
at which inferences are made from the GP model.  Any duplicated values
will be removed. These could be set at sample quantiles of \code{data}
using \code{\link[stats]{quantile}}.}

\item{n_v}{A numeric scalar.
The \code{n_v} largest values in \code{u_vec} will be used as
validation thresholds to quantify the predictive performance of the
GP models fitted at the thresholds in \code{u_vec}.}

\item{use_rcpp}{A logical scalar.  If TRUE (the default) use the
revdbayes function \code{\link[revdbayes]{rpost_rcpp}} for posterior
simulation.  Otherwise, we use \code{\link[revdbayes]{rpost}}.}

\item{...}{Further (optional) arguments to be passed to the
  \code{\link[revdbayes]{revdbayes}} function
  \code{\link[revdbayes]{rpost_rcpp}} (or \code{\link[revdbayes]{rpost}}).
  In particular:
\itemize{
  \item {\code{n}} {The size of the posterior sample used to perform
    predictive inference.  Default: \code{n = 1000}.}
  \item {\code{prior}} {A prior for the GP parameters, set using
    \code{\link[revdbayes]{set_prior}}.  Default: \code{prior = "flat"}
    with \code{min_xi = -1}.}
  \item {\code{h_prior}} {A list of further arguments (hyperparameters)
    for the GP prior specified in \code{prior}.}
  \item {\code{bin_prior}} {A prior for the threshold exceedance
    probability \eqn{p}, set using \code{\link[revdbayes]{set_bin_prior}}.
    Default: \code{prior = "jeffreys"}.}
  \item {\code{h_bin_prior}} {A list of further arguments (hyperparameters)
    for the binomial prior specified in \code{bin_prior}.}
}}
}
\value{
An object (list) of class \code{"thresh"}, containing the components
  \itemize{
    \item{\code{pred_perf}:} A numeric matrix with \code{length(u_vec)}
    rows and \code{n_v} columns.  Each column contains the values of
    the measure of predictive performance
    \item{\code{u_vec}:} The argument \code{u_vec} to \code{ithresh}.
    \item{\code{v_vec}:} A numeric vector.  The validation thresholds
      implied by the argument \code{n_v} to \code{ithresh}.
    \item{\code{u_ps}:} A numeric vector. The approximate levels of the
      sample quantiles to which the values in \code{u_vec} correspond,
      i.e. the approximate percentage of the data the lie at or below
      each element in \code{u_vec}.
    \item{\code{v_ps}:} A numeric vector.  The values in \code{u_ps}
      that correspond to the validation thresholds.
    \item{\code{data}:} The argument \code{data} to \code{ithresh}
      detailed above, with any missing values removed.
  }
}
\description{
Produces a diagnostic plot to assist in the selection of an extreme value
threshold in the case where the data can be treated as independent and
identically distributed (i.i.d.) observations.  For example, it could be
that these observations are the cluster maxima resulting from the
declustering of time series data.  The plot is based on the predictive
ability of the models fitted using each of a user-supplied set of
thresholds, assessed using leave-one-out cross-validation.
These models are based on a Generalized Pareto (GP) distribution for
threshold excesses and a binomial model for the probability of threshold
exceedance.  See
\href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)}
for details.
}
\details{
For a given threshold in \code{u_vec}:
\itemize{
  \item {the number of values in \code{data} that exceed the threshold,
    and the amounts (the \emph{threshold excesses}) by which these value
    exceed the threshold are calculated;}
  \item {\code{\link[revdbayes]{rpost_rcpp}}
    (or \code{\link[revdbayes]{rpost}}) is used to sample from the posterior
    distributions of the parameters of a GP model for the threshold
    excesses and a binomial model for the probability of threshold
    exceedance;}
  \item {the ability of this binomial-GP model to predict data
    thresholded at the threshold(s) specified by \code{n_v} is
    assessed using leave-one-out cross-validation (the measure of
    this is given in equation (7) of
    \href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)}.}
}
  See \href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)}
  and the introductory threshr vignette for further details and examples.
}
\examples{

# Gulf of Mexico significant wave heights
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
gom_cv <- ithresh(data = gom, u_vec = u_vec, n_v = 4)
plot(gom_cv, lwd = 2)

# North Sea significant wave heights
u_vec <- quantile(ns, probs = seq(0, 0.95, by = 0.05))
ns_cv <- ithresh(data = ns, u_vec = u_vec, n_v = 3,
  prior = "mdi", h_prior = list(a = 0.6))
plot(ns_cv, lwd = 2)
}
\references{
Northrop, P.J. and Attalides, N. (2016) Posterior propriety in
  Bayesian extreme value analyses using reference priors
  \emph{Statistica Sinica}, \strong{26}(2), 721--743
  \url{http://dx.doi.org/10.5705/ss.2014.034}.

Northrop, P. J., Attalides, N. and Jonathan, P. (2017)
  Cross-validatory extreme value threshold selection and uncertainty
  with application to ocean storm severity.
  \emph{Journal of the Royal Statistical Society Series C: Applied
  Statistics}, \strong{66}(1), 93-120.
  \url{http://dx.doi.org/10.1111/rssc.12159}
}
\seealso{
\code{\link{plot.thresh}} for the S3 plot method for objects of
  class \code{thresh}.

\code{\link[revdbayes]{rpost}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of the arguments
  that can be passed to
  \code{\link[revdbayes]{rpost_rcpp}}/\code{\link[revdbayes]{rpost}}.

\code{\link[revdbayes]{set_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for GP parameters.

\code{\link[revdbayes]{set_bin_prior}} in the
  \code{\link[revdbayes]{revdbayes}} package for details of how to set a
  prior distribution for the exceedance probability \eqn{p}.

\code{\link[stats]{quantile}}.
}
