% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_simulation.R
\name{plot.bc_sim_study}
\alias{plot.bc_sim_study}
\title{Plot method for objects of class "bc_sim_study"}
\usage{
\method{plot}{bc_sim_study}(x, sum = c("mean", "median", "quantiles"),
  probs = 0.5, which_lambdas = 1:length(x$bcthresh_args$lambda),
  legend_pos = "bottom", ...)
}
\arguments{
\item{x}{an object inheriting from class "bc_sim_study", a result of a call to
\code{\link{bc_sim_study}}.}

\item{sum}{Type of summary of the CV measures.}

\item{probs}{If \code{sum = "quantiles"} then \code{probs} provides
the argument to \code{\link[stats]{quantile}}.}

\item{which_lambdas}{A numeric vector.  Specifies which values of
\eqn{\lambda}, that is, the components of \code{x$lambda}, to use in the
plot.  The default is to use all these values.}

\item{legend_pos}{The position of the legend specified using the argument
\code{x} in \code{\link[graphics]{legend}}.}

\item{...}{Additional arguments to be passed to
\code{\link[graphics]{matplot}}.}
}
\value{
Nothing.
}
\description{
\code{plot} method for class "bc_sim_study".
}
\details{
Plots the measure of predictive performance against quantile
  of training threshold for different values of \eqn{\lambda}: those
  stored in \code{x$lambda}.
}
\section{Examples}{

See the examples in \code{\link{bc_sim_study}}.
}

