% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_simulation.R
\name{plot.bc_sim_study}
\alias{plot.bc_sim_study}
\title{Plot method for objects of class "bc_sim_study"}
\usage{
\method{plot}{bc_sim_study}(x, type = 1, stat = -1, which_N = 1,
  which_lambdas = 1:length(x$bcthresh_args$lambda), normalise = FALSE,
  legend_pos, summary = FALSE, digits = 2, ...)
}
\arguments{
\item{x}{an object inheriting from class "bc_sim_study", a result of a call to
\code{\link{bc_sim_study}}.}

\item{type}{An integer scalar.  Determine the type of plot.
\itemize{
  \item{type = 1: }{plots averages measures of predictive performance
    against threshold level by the value of \eqn{\lambda}.}
  \item{type = 2: }{plots, for each value of \eqn{\lambda}, the proportions
    of simulations on which each threshold has the largest measure of
    predictive performance.}
  \item{type = 3: }{plots kernel density estimates of the predictive
    density of the median of \eqn{N}-year maxima by the value of
    \eqn{\lambda}, using the best threshold for each \eqn{\lambda}, with the
    true value of the median indicated using a vertical line.}
}}

\item{stat}{A numeric scalar.  Determines the statistic(s) used to
summarise measures of performance across different simulated datasets,
for each each combination of \eqn{\lambda} and threshold level.
If \code{stat = -1} then the mean is used.  Otherwise, the
100\code{stat}\% sample quantile is used, with \code{stat} being passed
as the argument \code{probs} to \code{\link[stats]{quantile}}.
For example, if \code{stat = 0.5} then the sample median is used.}

\item{which_N}{A numeric scalar  Specifies which value of \eqn{N}, that is,
which component of \code{x$MN_args$N} is used as the time horizon when
\code{type = 2}.}

\item{which_lambdas}{A numeric vector.  Specifies which values of
\eqn{\lambda}, that is, the components of \code{x$bcthresh_args$lambda},
to include in the plot.  The default is to use all these values.}

\item{normalise}{A logical scalar.  Controls what is plotted on the
vertical axis of the plot.  If \code{normalise = FALSE} then the
measure of performance defined in equation (7) of Northrop et al. (2017)
is plotted. Otherwise, these measures are normalised in the manner of
equation (14) Northrop et al. (2017).  The same normalisation constant
is used across of values of \eqn{\lambda} that appear in the plot,
with the property that the values sum to 1 for one value of \eqn{\lambda}
and less than 1 for all other values of \eqn{\lambda}.}

\item{legend_pos}{The position of the legend that indicates which curves
relates to which value of \eqn{\lambda}, to be passed as the argument
\code{x} in \code{\link[graphics]{legend}}.  The default is
\code{"bottom"} for \code{type = 1} and \code{"topleft"} for
\code{type = 2}.}

\item{summary}{A logical scalar.  If \code{type = 2} should we add a legend,
in the top right of the plot, giving the estimated bias, standard
deviation and root mean squared error for each value of \eqn{lambda}?}

\item{digits}{An integer scalar.  The argument \code{digits} to be passed
to \code{\link[base:Round]{signif}} when rounding the
statistics produced when \code{summary = TRUE}.}

\item{...}{Additional graphical parameters to be passed to
\code{\link[graphics]{matplot}} and \code{\link[graphics]{legend}}.
The default setting plots solid lines (\code{lty = 1}) of width 2
(\code{lwd = 2}) using \code{col = 1:length(x$bcthresh_args$lambda)}.}
}
\value{
Nothing.
}
\description{
\code{plot} method for class "bc_sim_study". Plots the summaries of
predictive performance against quantile of training threshold for different
values of \eqn{\lambda}: those stored in \code{x$lambda}.
}
\details{
If \code{stat = -1} (the plot contains sample means) or if
  \code{stat} is close to zero (plot contains estimates of low quantiles)
  then it is possible that a curve on the plot may be incomplete.  This
  indicates that, for a particular \eqn{\lambda} and threshold level, a
  measure of predictive performance is \code{-Inf} for at least one
  simulated dataset. This occurs when an observation in the data lies above
  the estimated upper end point of the predictive distribution produced when
  this observation is removed.
}
\section{Examples}{

See the examples in \code{\link{bc_sim_study}}.
}

