% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthresh_methods.R
\name{plot.bcthresh}
\alias{plot.bcthresh}
\title{Plot method for objects of class "bcthresh"}
\usage{
\method{plot}{bcthresh}(x, y, which_v = 1, ...)
}
\arguments{
\item{x}{an object inheriting from class "bcthresh", a result of a call to
\code{\link{bcthresh}}.}

\item{y}{Not used.}

\item{which_v}{A numeric scalar.  Specifies the validation threshold, that
is the indices of the argument \code{v_vec} to \code{\link{bcthresh}},
to use in the plot.}

\item{...}{Additional arguments to be passed to
\code{\link[graphics]{matplot}}.}
}
\value{
Nothing.
}
\description{
\code{plot} method for class "bcthresh".
}
\details{
Add details.
}
