% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcthreshpred_methods.R
\name{plot.bcthreshpred}
\alias{plot.bcthreshpred}
\title{Plot diagnostics an bcthreshpred object}
\usage{
\method{plot}{bcthreshpred}(x, which_lambdas = 1:length(x$lambda),
  legend_pos, ...)
}
\arguments{
\item{x}{an object of class "bthreshpred", a result of a call to
\code{\link{ithresh}}.}

\item{which_lambdas}{A numeric vector.  Specifies which values of
\eqn{\lambda}, that is, the components of \code{x$lambda},
to include in the plot.  The default is to use all these values.}

\item{legend_pos}{The position of the legend specified using the argument
\code{x} in \code{\link[graphics]{legend}}.  If \code{legend_pos} is
missing then it is set internally, based on \code{x$type}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{matplot}}.}
}
\description{
\code{plot} method for class "bcthreshpred".  Produces plots to summarise
the predictive inferences made by \code{\link{predict.bcthresh}}.
}
