% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.ithresh}
\alias{plot.ithresh}
\title{Plot diagnostics an ithresh object}
\usage{
\method{plot}{ithresh}(x, y, ..., which_val = NULL, prob = TRUE,
  top_scale = TRUE, add_legend = FALSE, legend_pos = "topleft")
}
\arguments{
\item{x}{an object of class "ithresh", a result of a call to
\code{\link{ithresh}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to \code{\link[graphics]{matplot}}
and/or \code{\link[graphics]{legend}} and/or \code{\link[graphics]{axis}}.}

\item{which_val}{A numeric vector specifying the validation thresholds, that
is the indices of the argument \code{v_vec} to \code{\link{ithresh}}.}

\item{prob}{A logical scalar.  If \code{TRUE} then the levels of thresholds
are represented by the proportion of observations that lie below a
threshold.  If \code{prob = FALSE} then the values of the thresholds are
used.}

\item{top_scale}{A logical scalar indicating Whether or not to add a scale
to the top horizontal axis.  If this is added it gives the threshold on
the scale not chosen by \code{prob}.}

\item{add_legend}{A logical scalar indicating whether or not to add a
legend to the plot.  If \code{method = "cv"} then the legend gives the
levels of the validation thresholds.}

\item{legend_pos}{The position of the legend (if required) specified using
the argument \code{x} in \code{\link[graphics]{legend}}.}
}
\description{
\code{plot} method for class "ithresh".  Produces an extreme value
threshold diagnostic plot based on an analysis performed by
\code{\link{ithresh}}.
}
\details{
Produces plots of the \emph{threshold weights}, defined in
 equation (14) of
  \href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)},
  against training threshold.  A line is produced for each of the validation
  thresholds chosen in \code{which_v}.  The result is a plot like those in
  the top row of Figure 7 in
  \href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)}.
}
\examples{
\dontrun{
# [Smoother plots result from making n larger than the default n = 1000.]

u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
gom_cv <- ithresh(data = gom, u_vec = u_vec, n_v = 4)
plot(gom_cv, lwd = 2, add_legend = TRUE, legend_pos = "topleft")
mtext("significant wave height / m", side = 3, line = 2.5)
}
}
\seealso{
\code{\link{ithresh}} for threshold selection in the i.i.d. case
  based on leave-one-out cross-validaton.

\code{\link{summary.ithresh}} Summarizing measures of threshold
  predictive performance.
}
