% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predict.bcthresh}
\alias{predict.bcthresh}
\title{Predictive inference for the largest value observed in N years.}
\usage{
\method{predict}{bcthresh}(object, lambda = object$lambda, ...)
}
\arguments{
\item{object}{An object of class \code{"bcthresh"}, a result of a call to
\code{\link{bcthresh}}.}

\item{lambda}{A numeric vector containing values of the Box-Cox
transformation parameter \eqn{\lambda} to use when performing predictive
inference.  These must be contained in \code{object$lambda}.}

\item{...}{Additional arguments to be passed to
\code{\link{predict.ithresh}}.}
}
\description{
\code{predict} method for class \code{"bcthresh"}.  Calls
\code{\link{predict.ithresh}} for each value of \eqn{\lambda} in
\code{lambda}.
}
\examples{
# Set a prior: flat for GP parameters, Haldane for P(exceedance)
prior_args <- list(prior = "flatflat", bin_prior = "haldane",
                   h_prior = list(min_xi = -Inf))

## Gulf of Mexico significant wave heights ------------------

gprobs <- seq(0.1, 0.8, 0.1)
glambda <- seq(1, 3, 0.5)
gom_args <- list(data = gom, probs = gprobs, lambda = glambda)
gom_lambda <- do.call(bcthresh, c(gom_args, prior_args))
dgom <- predict(gom_lambda, type = "d")
}
