% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predict.ithresh}
\alias{predict.ithresh}
\title{Predictive distribution function of the largest value observed in N years.}
\usage{
\method{predict}{ithresh}(object, npy = NULL, n_years = 100,
  which_u = "best", which_v = 1, u_prior = NULL, type = c("p", "d", "q",
  "i", "r"), ...)
}
\arguments{
\item{object}{An object of class "ithresh", a result of a call to
\code{\link{ithresh}}.}

\item{npy}{A numeric scalar. The mean number of observations per year
of data, after excluding any missing values, i.e. the number of
non-missing observations divided by total number of years of non-missing
data.}

\item{n_years}{A numeric vector. Values of N.}

\item{which_u}{Either a character scalar or a numeric vector.

  If \code{which_u} is a character scalar it must be either "best" or "all".
  If \code{which_u = "best"} then the threshold achieving the largest
  measure of predictive performance in \code{object$pred_perf}, based
  on the validation threshold selected using \code{n_v}, is used to
  perform prediction.
  If \code{which_u = "all"} then \emph{all} the thresholds are used to
  perform prediction.  The inferences from each threshold are weighted
  according to equation (15) in
  \href{https://doi.org/10.1111/rssc.12159}{Northrop et al. (2017)}
  based on the prior probabilities of thresholds in \code{u_prior}
  and column \code{n_v} of the measures of predictive performance in
  \code{object$pred_perf}.

  Otherwise, \code{which_u} is a numeric vector that indicates which
  elements of \code{u_vec} the user wishes to select as thresholds on
  which to base prediction, that is, it must contain integers in
  {1, ..., \code{length(object$u_vec)}}.}

\item{which_v}{A numeric scalar. Indicates which element of
\code{object$v_vec} is used in selecting a single threshold
(if \code{which_u = "best"}) or weighting the inferences from
all thresholds (if \code{which_u = "all"}).}

\item{u_prior}{A numeric vector.  Prior probabilities for the thresholds
in \code{u_prior}.  By default this is set to a vector of length
\code{length(u_vec)} with each element equal to 1/\code{length(u_vec)}.}

\item{type}{A character vector.  Indicates which type of inference is
  required:
\itemize{
  \item "p" for the predictive distribution function,
  \item "d" for the predictive density function,
  \item "q" for the predictive quantile function,
  \item "i" for predictive intervals,
  \item "r" for random generation from the predictive distribution.
}}

\item{...}{Additional arguments passed on to
\code{\link[revdbayes]{predict.evpost}}, in particular \code{x}.}
}
\value{
An object of class "pred_ithresh"
}
\description{
\code{predict} method for class "ithresh".
}
\details{
Add details
}
\examples{
# Gulf of Mexico significant wave heights, default priors.
u_vec <- quantile(gom, probs = seq(0, 0.95, by = 0.05))
npy_gom <- length(gom)/105
gom_cv <- ithresh(data = gom, u_vec = u_vec, n_v = 4)
pjn <- predict(gom_cv, npy = npy_gom)
}
\references{
Northrop, P. J., Attalides, N. and Jonathan, P. (2017)
  Cross-validatory extreme value threshold selection and uncertainty
  with application to ocean storm severity.
  \emph{Journal of the Royal Statistical Society Series C: Applied
  Statistics}, \strong{66}(1), 93-120.
  \url{http://dx.doi.org/10.1111/rssc.12159}
}
