% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{print.ithresh}
\alias{print.ithresh}
\title{Print method for objects of class "ithresh"}
\usage{
\method{print}{ithresh}(x, digits = 2, ...)
}
\arguments{
\item{x}{an object inheriting from class "ithresh", a result of a call to
\code{\link{ithresh}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{format}} and \code{\link[base:Round]{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class "ithresh".
}
\details{
Prints the call and a matrix of the estimates threshold weights.
  Each row gives the weights for each training threshold for a given
  validation threshold.  The row and column names are the approximate
  quantile levels of the thresholds.
}
\seealso{
\code{\link{gevreg}} GEV generalized linear regression modelling.
}
