% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshr-internal.R
\name{threshr-internal}
\alias{threshr-internal}
\alias{bc_gm}
\alias{inv_bc}
\title{Internal threshr functions}
\usage{
bc_gm(x, lambda = 1, lngm = 0, lambda_tol = 1/50, m = 4)

inv_bc(x, lambda, lambda_tol = 1/50, m = 4)
}
\description{
Internal threshr functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
